+++ py/dml/compat.py
@@ -374,3 +374,30 @@ class dml12_int(CompatFeature):
     '''
     short = "Use legacy integer semantics in DML 1.2"
     last_api_version = api_6
+
+
+@feature
+class experimental_vect(CompatFeature):
+    '''<a id="experimental_vect"/> This compat feature
+    controls how DMLC reacts to uses of the `vect` syntax in files
+    where the [`simics_util_vect` provisional feature](provisional-auto.html#simics_util_vect)
+    is not enabled.
+
+    When the `experimental_vect` compatibility feature is
+    enabled, such uses are permitted, and give a `WEXPERIMENTAL`
+    warning in DML 1.4 (but no warning in DML 1.2). When
+    `experimental_vect` is disabled, DMLC forbids the `vect`
+    syntax.
+
+    '''
+    short = "Permit vect syntax without 'provisional simics_util_vect'"
+    last_api_version = api_7
+
+
+@feature
+class warning_statement(CompatFeature):
+    '''This compatibility feature enables the `_warning` statement.
+    This turned out to not be very useful, so in Simics API 8 and
+    newer the feature is no longer allowed.'''
+    short = "Allow the `_warning` statement"
+    last_api_version = api_7