+++ test/tests.py
@@ -1452,6 +1452,7 @@ all_tests.append(PortingConvert(
        ('utility.dml', None, 'PRENAME_TEMPLATE'),
        ('utility.dml', None, 'PABSTRACT_TEMPLATE'),
        ('utility.dml', None, 'PHASH'),
+       ('utility.dml', None, 'PHASHELSE'),
        ('utility.dml', None, 'PIFAND'),
        ('utility.dml', None, 'PANDOR'),
        ('utility.dml', None, 'PCHANGE_INARGS'),